Language.Scheme.Compiler.Libraries

Description

This module contains support for compiling libraries of scheme code.

Documentation

importAll

Arguments

:: Env
Compilation environment

-> Env
Compilation meta environment, containing code from modules.scm

-> [LispVal]
Modules to import

-> CompLibOpts
Misc options required by compiler library functions

-> CompOpts
Misc options required by compiler functions

-> IOThrowsError [HaskAST]
Compiled code

Import all given modules and generate code for them