Language.Scheme.Plugins.CPUTime

Description

This module wraps System.CPUTime so that it can be used directly by Scheme code.

More importantly, it serves as an example of how to wrap existing Haskell code so that it can be loaded and called by husk.

See 'examples/ffi/ffi-cputime.scm' in the husk source tree for an example of how to call into this module from Scheme code.

Documentation

get :: [LispVal] -> IOThrowsError LispVal

Wrapper for CPUTime.getCPUTime

precision :: [LispVal] -> IOThrowsError LispVal

Wrapper for CPUTime.cpuTimePrecision