Language.Scheme.Compiler

Description

This module contains an experimental Scheme to Haskell compiler.

The compiler performs the following transformations:

 Scheme AST (LispVal) => Haskell AST (HaskAST) => Compiled Code (String)

The GHC compiler is then used to create a native executable.

At present, the focus has just been on creating a compiler that will generate correct, working code. Many optimizations could and need to be made for time and space...

Documentation

data CompOpts

Constructors

CompileOptions

Fields

coptsThisFunc :: String

coptsThisFuncUseValue :: Bool

coptsThisFuncUseArgs :: Bool

coptsNextFunc :: Maybe String

defaultCompileOptions :: String -> CompOpts

createAstFunc :: CompOpts -> [HaskAST] -> HaskAST

createAstCont :: CompOpts -> String -> String -> HaskAST

data HaskAST

Constructors

AstAssignM String HaskAST

AstFunction

Fields

astfName :: String

astfArgs :: String

astfCode :: [HaskAST]

AstValue String

AstContinuation

Fields

astcNext :: String

astcArgs :: String

showValAST :: HaskAST -> String

joinL :: forall a. [[a]] -> [a] -> [a]

astToHaskellStr :: LispVal -> String

headerModule :: [String]

headerImports :: [String]

header :: [String]

initializeCompiler :: Env -> IOThrowsError [HaskAST]

compileLisp :: Env -> String -> String -> Maybe String -> IOThrowsError [HaskAST]

compileBlock :: String -> Maybe String -> Env -> [HaskAST] -> [LispVal] -> IOThrowsError [HaskAST]

compileScalar :: String -> CompOpts -> IOThrowsError [HaskAST]

compileLambdaList :: [LispVal] -> IOThrowsError String

compile :: Env -> LispVal -> CompOpts -> IOThrowsError [HaskAST]

mcompile :: Env -> LispVal -> CompOpts -> IOThrowsError [HaskAST]

mfunc :: Env -> LispVal -> (Env -> LispVal -> CompOpts -> IOThrowsError [HaskAST]) -> CompOpts -> IOThrowsError [HaskAST]

compileSpecialFormEntryPoint :: String -> String -> CompOpts -> IOThrowsError HaskAST

compileSpecialForm :: String -> String -> CompOpts -> IOThrowsError HaskAST

compileExpr :: Env -> LispVal -> String -> Maybe String -> IOThrowsError [HaskAST]

compileApply :: Env -> LispVal -> CompOpts -> IOThrowsError [HaskAST]

Compiles each argument to a function call, and then uses apply to call the function