Language.Scheme.Parser

Description

This module implements parsing of Scheme code.

Documentation

lispDef :: LanguageDef ()

Language definition for Scheme

Higher level parsing

mainParser :: Parser LispVal

readOrThrow :: Parser a -> String -> ThrowsError a

Use a parser to parse the given text, throwing an error if there is a problem parsing the text.

readExpr :: String -> ThrowsError LispVal

Parse an expression from a string of text

readExprList :: String -> ThrowsError [LispVal]

Parse many expressions from a string of text

Low level parsing

symbol :: Parser Char

parseExpr :: Parser LispVal

Parse an expression

parseAtom :: Parser LispVal

parseBool :: Parser LispVal

parseChar :: Parser LispVal

parseOctalNumber :: Parser LispVal

parseBinaryNumber :: Parser LispVal

parseHexNumber :: Parser LispVal

parseDecimalNumber :: Parser LispVal

Parser for Integer, base 10

parseNumber :: Parser LispVal

Parse an integer in any base

parseRealNumber :: Parser LispVal

Parse a floating point number

parseRationalNumber :: Parser LispVal

parseComplexNumber :: Parser LispVal

parseEscapedChar :: forall st. GenParser Char st Char

parseString :: Parser LispVal

parseVector :: Parser LispVal

parseList :: Parser LispVal

parseDottedList :: Parser LispVal

parseQuoted :: Parser LispVal

parseQuasiQuoted :: Parser LispVal

parseUnquoted :: Parser LispVal

parseUnquoteSpliced :: Parser LispVal